+++ hutool-cron/src/test/java/cn/hutool/cron/pattern/CronPatternNextMatchTest.java
@@ -14,7 +14,7 @@ public class CronPatternNextMatchTest {
 		CronPattern pattern = new CronPattern("23 12 * 12 * * *");
 		//noinspection ConstantConditions
 		final Calendar calendar = pattern.nextMatchAfter(
-				DateUtil.parse("2022-04-12 09:12:12").toCalendar());
+				DateUtil.parse("2022-04-12 09:12:24").toCalendar());
 
 		Console.log(DateUtil.date(calendar));
 		Assert.assertTrue(pattern.match(calendar, true));